refactor: split test doubles into TestDoubles/ folder
