+++ server/tests/Dotbot.Server.Tests/Integration/TemplatesApiFactory.cs
@@ -1,6 +1,5 @@
-using Dotbot.Server.Models;
 using Dotbot.Server.Services;
-using Microsoft.Agents.Core.Models;
+using Dotbot.Server.Tests.Integration.TestDoubles;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Mvc.Testing;
 using Microsoft.Extensions.Configuration;
@@ -59,36 +58,4 @@ public sealed class TemplatesApiFactory : WebApplicationFactory<Program>
             services.AddSingleton<IConversationReferenceStore>(new NullConversationReferenceStore());
         });
     }
-}
-
-public sealed class InMemoryTemplateStorage : ITemplateStorageService
-{
-    private readonly List<QuestionTemplate> _saved = [];
-
-    public IReadOnlyList<QuestionTemplate> Saved => _saved;
-
-    public Task SaveTemplateAsync(QuestionTemplate template)
-    {
-        _saved.Add(template);
-        return Task.CompletedTask;
-    }
-
-    public Task<QuestionTemplate?> GetTemplateAsync(string projectId, Guid questionId, int version)
-        => Task.FromResult(_saved.FirstOrDefault(x =>
-            x.Project.ProjectId == projectId
-            && x.QuestionId == questionId
-            && x.Version == version));
-}
-
-internal sealed class NullAdministratorService : IAdministratorService
-{
-    public Task SeedIfEmptyAsync() => Task.CompletedTask;
-    public Task<bool> IsAdministratorAsync(string email) => Task.FromResult(false);
-}
-
-internal sealed class NullConversationReferenceStore : IConversationReferenceStore
-{
-    public Task LoadAsync() => Task.CompletedTask;
-    public void AddOrUpdate(string userObjectId, ConversationReference reference) { }
-    public ConversationReference? Get(string userObjectId) => null;
 }
